Add new entities and config files for TDD project
This commit introduces multiple new entities: 'Costumer', 'Car', 'CarCategory', 'BaseRepository', and 'Tax'. It also includes necessary configurations for the Test-Driven Development (TDD) project. The '.nycrc.json' is added for nyc configuration. A 'seed/index.js' is added for seeding the database. The 'package-lock.json' file is modified as a result of installing new dependencies.

These changes are necessary to set up the structure of the project and manage the testing environment. It's part of the ongoing process of developing the project using TDD methodology.